Move libusb support into a shared location
